add realm, pagination, settings
