--- MovieDB/MovieDB/Common/Persistance/FavoriteMovie+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  FavoriteMovie+CoreDataClass.swift
-//  MovieDB
-//
-//  Created by Dmytro Vasylenko on 24.02.2025.
-//
-//
-
-import Foundation
-import CoreData
-
-@objc(FavoriteMovie)
-public class FavoriteMovie: NSManagedObject {
-
-}
