+++ MovieDB/MovieDB/Modules/FavoritesModule/FaforitesPresenter.swift
@@ -1,0 +1,8 @@
+//
+//  FaforitesPresenter.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 12.03.2025.
+//
+
+import Foundation
