+++ MovieDB/MovieDB/Modules/FavoritesModule/FavoritesBuilder.swift
@@ -1,0 +1,8 @@
+//
+//  FavoritesBuilder.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 12.03.2025.
+//
+
+import Foundation
