+++ MovieDB/MovieDB/Modules/FavoritesModule/FavoritesInteractor.swift
@@ -1,0 +1,8 @@
+//
+//  FavoritesInteractor.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 12.03.2025.
+//
+
+import Foundation
