+++ MovieDB/MovieDB/Modules/FavoritesModule/FavoritesRouter.swift
@@ -1,0 +1,8 @@
+//
+//  FavoritesRouter.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 12.03.2025.
+//
+
+import Foundation
