--- MovieDB/MovieDB/Modules/MovieDetail/Model/Genre.swift
@@ -1,13 +1,0 @@
-//
-//  Genre.swift
-//  MovieDB
-//
-//  Created by Dmytro Vasylenko on 24.02.2025.
-//
-
-import Foundation
-
-struct Genre: Decodable {
-    let id: Int
-    let name: String
-}
