--- MovieDB/MovieDB/Modules/MovieDetail/Model/MovieDetail.swift
@@ -1,50 +1,0 @@
-//
-//  MovieDetail.swift
-//  MovieDB
-//
-//  Created by Dmytro Vasylenko on 25.02.2025.
-//
-
-import Foundation
-
-struct MovieDetail: Decodable {
-    let id: Int
-    let title: String
-    let overview: String
-    let posterPath: String?
-    let backdropPath: String?
-    let voteAverage: Double
-    let releaseDate: String
-    let genres: [Genre]?
-    
-    enum CodingKeys: String, CodingKey {
-        case id, title, overview
-        case posterPath = "poster_path"
-        case backdropPath = "backdrop_path"
-        case voteAverage = "vote_average"
-        case releaseDate = "release_date"
-        case genres
-    }
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        id = try container.decode(Int.self, forKey: .id)
-        title = try container.decode(String.self, forKey: .title)
-        overview = try container.decodeIfPresent(String.self, forKey: .overview) ?? "No description available"
-        posterPath = try container.decodeIfPresent(String.self, forKey: .posterPath)
-        backdropPath = try container.decodeIfPresent(String.self, forKey: .backdropPath)
-        voteAverage = try container.decodeIfPresent(Double.self, forKey: .voteAverage) ?? 0.0
-        releaseDate = try container.decodeIfPresent(String.self, forKey: .releaseDate) ?? "Unknown"
-        genres = try container.decodeIfPresent([Genre].self, forKey: .genres) ?? []
-    }
-    
-    var posterURL: URL? {
-        guard let path = posterPath else { return nil }
-        return URL(string: "https://image.tmdb.org/t/p/w500\(path)")
-    }
-    
-    var backdropURL: URL? {
-        guard let path = backdropPath else { return nil }
-        return URL(string: "https://image.tmdb.org/t/p/w780\(path)")
-    }
-}
