+++ MovieDB/MovieDB/Modules/MovieDetail/MovieDetailPresenter.swift
@@ -1,0 +1,8 @@
+//
+//  MovieDetailPresenter.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 12.03.2025.
+//
+
+import Foundation
