+++ MovieDB/MovieDB/Modules/MovieDetail/MovieDetailRouter.swift
@@ -1,0 +1,8 @@
+//
+//  MovieDetailRouter.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 12.03.2025.
+//
+
+import Foundation
