+++ MovieDB/MovieDB/Modules/MovieDetail/MoviesDetailInteractor.swift
@@ -1,0 +1,8 @@
+//
+//  MoviesDetailInteractor.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 12.03.2025.
+//
+
+import Foundation
