+++ MovieDB/MovieDB/Modules/SettingsModule/SettingsRouter.swift
@@ -1,0 +1,8 @@
+//
+//  SettingsRouter.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 16.03.2025.
+//
+
+import Foundation
