+++ MovieDB/MovieDB/Modules/SettingsModule/SettingsViewController.swift
@@ -1,0 +1,8 @@
+//
+//  SettingsViewController.swift
+//  MovieDB
+//
+//  Created by Dmytro Vasylenko on 16.03.2025.
+//
+
+import Foundation
